Rename root deduplicator to delegate deduplicator since it is delegating
