+++ data/deduplicator/delegate/delegate.go
@@ -1,4 +1,4 @@
-package root
+package delegate
 
 /* CHECKLIST
  * [ ] Uses interfaces as appropriate
