+++ data/deduplicator/delegate/delegate_test.go
@@ -1,9 +1,9 @@
-package root_test
+package delegate_test
 
 import (
 	. "github.com/onsi/ginkgo"
 	// . "github.com/onsi/gomega"
 )
 
-var _ = PDescribe("Root", func() {
+var _ = PDescribe("Delegate", func() {
 })
